+++ src/main/java/com/safetynetalerts/safetynet/controller/MedicalRecordController.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -60,6 +61,14 @@ public class MedicalRecordController {
 	public ResponseEntity<Void> addMedicalRecord(@RequestBody MedicalRecordDTO medicalRecordDTO) {
 		logger.debug("Adding medical record: {}", medicalRecordDTO);
 		
+		if(
+				medicalRecordDTO.getFirstName() == null || medicalRecordDTO.getFirstName().isEmpty() ||
+				medicalRecordDTO.getLastName() == null || medicalRecordDTO.getLastName().isEmpty()
+			) {
+				logger.error("First name or last name is missing");
+				return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+		}
+		
 		try {
 			medicalRecordService.addMedicalRecord(medicalRecordDTO);
 			logger.info("Added medical records successfully. {}", medicalRecordDTO);
@@ -67,7 +76,7 @@ public class MedicalRecordController {
 			
 		} catch (Exception e) {
 			logger.error("Error adding medical records: ", medicalRecordDTO ,e);
-			return ResponseEntity.status(500).build();
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
 		}
 	}
 	
@@ -80,6 +89,14 @@ public class MedicalRecordController {
 	public ResponseEntity<Void> updateMedicalRecord(@RequestBody MedicalRecordDTO medicalRecordDTO) {
 		logger.debug("Updating medical record: {}", medicalRecordDTO);
 		
+		if(
+				medicalRecordDTO.getFirstName() == null || medicalRecordDTO.getFirstName().isEmpty() ||
+				medicalRecordDTO.getLastName() == null || medicalRecordDTO.getLastName().isEmpty()
+			) {
+				logger.error("First name or last name is missing");
+				return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+		}
+		
 		try {
 			medicalRecordService.updatePerson(medicalRecordDTO);
 			logger.info("Updated medical record successfully: {}", medicalRecordDTO);
@@ -87,7 +104,7 @@ public class MedicalRecordController {
 			
 		} catch (Exception e) {
 			logger.error("Error updating medical record: ", medicalRecordDTO ,e);
-			return ResponseEntity.status(500).build();
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
 		}
 	}
 	
@@ -101,6 +118,13 @@ public class MedicalRecordController {
 	public ResponseEntity<Void> deleteMedicalRecord(@RequestParam String firstName, @RequestParam String lastName){
 		logger.debug("Deleting medical record with firstName: {} and lastName: {}" , firstName, lastName);
 		
+		if(firstName == null || firstName.isEmpty() ||
+			lastName == null || lastName.isEmpty()
+			) {
+				logger.error("First name or last name is missing");
+				return ResponseEntity.status(HttpStatus.BAD_REQUEST).build();
+		}
+		
 		try {
 			medicalRecordService.deleteMedicalRecord(firstName, lastName);
 			logger.info("Deleted medical record successfully with firstName: {} and lastName: {}" , firstName, lastName);
@@ -108,7 +132,7 @@ public class MedicalRecordController {
 			
 		} catch (Exception e) {
 			logger.error("Error deleting medical record with firstName: {} and lastName: {}" , firstName, lastName, e);
-			return ResponseEntity.status(500).build();
+			return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
 		}
 	}
